Changed ProductDetail Page

+++ app/src/main/java/com/droid/cleanarchitecture/db/ProductsEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "Products")
 data class ProductsEntity(
-    @PrimaryKey @ColumnInfo var productId: Int,
+    @PrimaryKey @ColumnInfo var productId: Long,
     @ColumnInfo var productName: String,
     @ColumnInfo var category: String,
     @ColumnInfo var price: String,
